+++ src/main/java/br/com/fiap/entities/Usuario.java
@@ -2,7 +2,7 @@ package br.com.fiap.entities;
 
 public class Usuario {
 
-    private int id;
+    private Long id;
     private String nome;
     private String email;
     private String senha;
@@ -10,18 +10,18 @@ public class Usuario {
     public Usuario(){
     }
 
-    public Usuario(int id, String nome, String email, String senha) {
+    public Usuario(Long id, String nome, String email, String senha) {
         this.id = id;
         this.nome = nome;
         this.email = email;
         this.senha = senha;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
